+++ tests/conftest.py
@@ -1,0 +1,5 @@
+import asyncio
+
+from rdp.api.main import startup_event
+
+asyncio.run(startup_event())
